% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFunctions.R
\name{BackwardR}
\alias{BackwardR}
\title{Backward step}
\usage{
BackwardR(Fpr, trsVec, nb.states)
}
\arguments{
\item{Fpr}{Fpr.}

\item{trsVec}{a vector of state transition probabilities.}

\item{nb.states}{an integer specifying the numbers of states.}
}
\description{
Backward step
}
\keyword{internal}
