% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncAux.R
\name{clusterseg}
\alias{clusterseg}
\title{Summarize the results of the coupled HMM.}
\usage{
clusterseg(x)
}
\arguments{
\item{x}{a matrix of status. Columns corresponds to series (individuals).}
}
\value{
a data.frame with 4 columns
 \describe{
\item{\code{sample}}{name of the sample (series). }
\item{\code{posbegin}}{beginning position.}
\item{\code{posend}}{ending position.}
\item{\code{status}}{status.}
}
}
\description{
Summarize the results of the coupled HMM.
}
