% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncEmis.R
\name{Emis.Gauss}
\alias{Emis.Gauss}
\title{Emis.Gauss}
\usage{
Emis.Gauss(X, esAvg, esVar)
}
\arguments{
\item{X}{data matrix of observations.}

\item{esAvg}{a numeric of the estimated mean for each state.}

\item{esVar}{a numeric of the estimated variance for each state.}
}
\description{
Emis.Gauss
}
\keyword{internal}
