% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncEmis.R
\name{Emis.LRR}
\alias{Emis.LRR}
\title{Emis.LRR}
\usage{
Emis.LRR(X, esAvg, esVar, weight = 0.05)
}
\arguments{
\item{X}{data matrix of observations.}

\item{esAvg}{a numeric of the estimated mean for each state.}

\item{esVar}{a numeric of the estimated variance for each state.}

\item{weight}{weight.}
}
\description{
Emis.LRR
}
\keyword{internal}
