% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncEmis.R
\name{EmisGb.Gauss}
\alias{EmisGb.Gauss}
\title{EmisGb.Gauss}
\usage{
EmisGb.Gauss(X, esAvgGb, esVarGb)
}
\arguments{
\item{X}{data matrix of observations.}

\item{esAvgGb}{a matrix of \code{nbK} rows and \code{nbI} columns of estimated mean.}

\item{esVarGb}{a matrix of \code{nbK} rows and \code{nbI} columns of estimated variance.}
}
\description{
EmisGb.Gauss
}
\keyword{internal}
