% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncAux.R
\name{ExpGrid}
\alias{ExpGrid}
\title{Wl Internal function calculating Wl.}
\usage{
ExpGrid(nb.states, nbI)
}
\arguments{
\item{nb.states}{an integer specifying the numbers of states.}

\item{nbI}{an integer specifying the number of series.}
}
\value{
a matrix containing all combination of possible state for nbI series.
}
\description{
Wl Internal function calculating Wl.
}
\keyword{internal}
