% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFunctions.R
\name{ForwardR}
\alias{ForwardR}
\title{Forward step}
\usage{
ForwardR(emisVec, initPr, trsVec)
}
\arguments{
\item{emisVec}{a vector of emission probabilities.}

\item{initPr}{a vector specifying initial state probabilities.}

\item{trsVec}{a vector of state transition probabilities.}
}
\description{
Forward step
}
\keyword{internal}
