% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toy example -  observations for 5 correlated samples.}
\format{A simulated matrix with 1000 rows and 5 columns. Each column is a series}
\description{
A matrix containing the observations for the 1,000 positions of 5
 correlated samples.
}
\examples{
data(toyexample)
# Variational inference of a coupled hidden Markov Chains
resCHMM <- coupledHMM(X = toydata, nb.states = 3, S = cor(toystatus),
                      omega.list = c(0.3, 0.5, 0.7, 0.9))
# Breakpoints positions and status of segments
info <- clusterseg(resCHMM$status)
# head(info)
}
\keyword{datasets}
