% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toystatus}
\alias{toystatus}
\title{Toy example -  status for 5 correlated samples.}
\format{A matrix of the hidden status corresponding to the \code{toydata} matrix.}
\description{
A matrix containing the hidden status for the 1,000 positions of 5
 correlated samples.
}
\keyword{datasets}
