% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi_algo.R
\name{viterbi_algo}
\alias{viterbi_algo}
\title{Implementation of the Viterbi algorithm}
\usage{
viterbi_algo(emisPr, transPr, initPr)
}
\arguments{
\item{emisPr}{a matrix of emission probabilities for the considering series.}

\item{transPr}{a matrix of state transition probabilities.}

\item{initPr}{a vector specifying initial state probabilities.}
}
\value{
\item{path}{the most likely path (state sequence).}
}
\description{
Implementation of the Viterbi algorithm
}
\keyword{internal}
