% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncAux.R
\name{Wl}
\alias{Wl}
\title{Wl Internal function calculating Wl.}
\usage{
Wl(ID.K, S, omega)
}
\arguments{
\item{ID.K}{as.matrix(expand.grid(list(c(1:nbI),c(1:nb.states))).}

\item{S}{a matrix of similarities between individuals.}

\item{omega}{.}
}
\value{
Wl.
}
\description{
Wl Internal function calculating Wl.
}
\keyword{internal}
